fix: Update fxa-relier-client to redirect off of the oauth server.
The relier client needs to know both the oauth and content server hosts.

The newest fxa-relier-client bounces off the oauth server instead of going directly to the content server. To handle this, we have to pass both the content server URI and oauth server URI to the relier-client on startup.